backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
networkx==3.1
semver==3.0.1
cyclonedx-bom==3.11.0
python-dateutil==2.8.2
gunicorn==21.2.0  # For production